requirements.txt
asgiref==3.7.2
Brotli==1.1.0
Django==5.0.1
django-configurations==2.5.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
gunicorn==21.2.0
packaging==23.2
typing_extensions>=4.11
whitenoise==6.6.0

# Essential dependencies for EditEngine business logic
celery[redis]>=5.2.0
redis>=4.5.0
cryptography>=41.0.0
langchain>=0.3.20
langchain-core>=0.3.60
langchain-openai>=0.3.20
langchain-anthropic>=0.3.10
langchain-google-genai>=2.1.0
langchain-mistralai>=0.2.10
langchain-perplexity>=0.1.0
langchain-text-splitters>=0.3.0
openai>=1.90.0
wikitextparser==0.56.4
pytest==8.4.0
pytest-django==4.11.1
pytest-cov==6.0.0
syrupy==4.6.1
coverage==7.8.2
python-dotenv==1.1.0
django-ratelimit>=4.0.0
